third_party/blink/web_tests/external/wpt/service-workers/service-worker/fetch-canvas-tainting-video-with-range-request.https-expected.txt

This is a testharness.js-based test.
[FAIL] synth and same-origin fallback range responses
  assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
Harness: the test ran to completion.
